visualaccess.scripts/requirements.txt
face-recognition==1.3.0
Pillow==10.1.0
Flask==3.0.0
Flask-Cors==4.0.0
